Stop users from deleting existing nodes in a model when in lockdown mode [#159145561]

+++ src/code/mixins/app-view.coffee
@@ -3,6 +3,7 @@ CodapStore      = require "../stores/codap-store"
 GoogleFileStore = require "../stores/google-file-store"
 HashParams      = require "../utils/hash-parameters"
 tr              = require '../utils/translate'
+AppSettingsStore = require '../stores/app-settings-store'
 
 module.exports =
 
@@ -26,7 +27,8 @@ module.exports =
         # 8 is backspace, 46 is delete
         if e.which in [8, 46] and not $(e.target).is('input, textarea')
           e.preventDefault()
-          deleteFunction()
+          if not AppSettingsStore.store.settings.lockdown
+            deleteFunction()
     else
       $(window).off 'keydown'
